+++ front/components.d.ts
@@ -81,6 +81,7 @@ declare module 'vue' {
     MapFiltersStatus: typeof import('./src/components/map/panels/MapFiltersStatus.vue')['default']
     MapGeocoder: typeof import('./src/components/map/MapGeocoder.vue')['default']
     MapLayerSwitcher: typeof import('./src/components/layerSwitcher/MapLayerSwitcher.vue')['default']
+    MapLayerToggles: typeof import('./src/components/map/MapLayerToggles.vue')['default']
     MapLegend: typeof import('./src/components/map/legend/MapLegend.vue')['default']
     MapQpvToggleButton: typeof import('./src/components/map/panels/MapQpvToggleButton.vue')['default']
     MapScoresDrawer: typeof import('./src/components/map/panels/MapScoresDrawer.vue')['default']
@@ -90,7 +91,6 @@ declare module 'vue' {
     MapSidePanelHeader: typeof import('./src/components/map/panels/sidepanel/MapSidePanelHeader.vue')['default']
     MapStyleOption: typeof import('./src/components/map/controls/MapStyleOption.vue')['default']
     MapSwitcher: typeof import('./src/components/mapSwitcher/MapSwitcher.vue')['default']
-    MapTopRightPanel: typeof import('./src/components/map/MapTopRightPanel.vue')['default']
     NavbarComponent: typeof import('./src/components/navbar/NavbarComponent.vue')['default']
     PermeabilityWidget: typeof import('./src/components/dashboard/widgets/PermeabilityWidget.vue')['default']
     PlantabilityCityCard: typeof import('./src/components/division/PlantabilityCityCard.vue')['default']
